[OWL-611] `data.go`->`marshal.go`
Renamed `data.go` as `marshal.go`. Refactoring and refinement of the
code for marshalling the JSON parameters were also done.
